+++ core/src/main/java/react4j/core/NativeAdapterComponent.java
@@ -65,7 +65,7 @@ public abstract class NativeAdapterComponent<
     }
     try
     {
-      _component.performComponentDidConstruct( props, context );
+      _component.performComponentDidConstruct();
     }
     finally
     {